+++ packages/locator/src/index.tsx
@@ -10,7 +10,6 @@ import {
 	Show,
 } from "solid-js"
 import { Portal } from "solid-js/web"
-import { Motion, Presence } from "@motionone/solid"
 import { isProd } from "@solid-primitives/utils"
 import { createElementBounds } from "@solid-primitives/bounds"
 import { makeEventListener } from "@solid-primitives/event-listener"
@@ -22,6 +21,7 @@ import { makeHoverElementListener } from "./hoverElement"
 import { createElementCursor } from "./elementCursor"
 import { openCodeSource, TargetIDE, TargetURLFunction } from "./goToSource"
 import { makeHoldKeyListener } from "./holdKeyListener"
+import { animate } from "motion"
 
 // TODO: contribute to solid-primitives
 const stopPropagation =
@@ -147,6 +147,7 @@ const ElementOverlay: Component<{
 	const width = createMemo<number>(prev => (props.width === null ? prev : props.width), 0)
 	const height = createMemo<number>(prev => (props.height === null ? prev : props.height), 0)
 	const transform = createMemo(() => `translate(${Math.round(left())}px, ${Math.round(top())}px)`)
+	const placeOnTop = createMemo(() => top() > window.innerHeight / 2)
 
 	return (
 		<div
@@ -161,22 +162,41 @@ const ElementOverlay: Component<{
 			<div
 				class={tw`absolute -inset-2 rounded border-2 border-cyan-800 border-opacity-80 bg-cyan-800 bg-opacity-30`}
 			/>
-			<Presence>
-				<Show when={!!props.name}>
-					<Motion.div
-						initial={{ opacity: 0, y: -10 }}
-						animate={{ opacity: 1, y: 0 }}
-						exit={{ opacity: 0, y: 10 }}
-						class={tw`absolute top-full inset-x-0 flex justify-center`}
+			<Show when={!!props.name}>
+				<div
+					class={tw`absolute ${
+						placeOnTop() ? "bottom-full" : "top-full"
+					} inset-x-0 flex justify-center`}
+				>
+					<div
+						class={tw`relative my-3 py-1 px-2`}
+						ref={el => {
+							let prevY = 0
+							createComputed(
+								on(placeOnTop, () => (prevY = el.getBoundingClientRect().top), { defer: true }),
+							)
+							createEffect(
+								on(
+									placeOnTop,
+									() => {
+										const currY = el.getBoundingClientRect().top
+										animate(el, { y: [prevY - currY, 0] }, { duration: 0.15 })
+									},
+									{ defer: true },
+								),
+							)
+						}}
 					>
-						<div class={tw`relative mt-3 py-1 px-2 bg-cyan-900 bg-opacity-80 rounded`}>
-							<span class={tw`text-cyan-50 font-mono text-sm leading-3`}>
-								{props.name}: <span class={tw`text-cyan-200`}>{props.tag}</span>
-							</span>
+						<div class={tw`absolute inset-0 bg-cyan-900 bg-opacity-80 rounded`}></div>
+						<div class={tw`absolute text-cyan-50 font-mono text-sm leading-4`}>
+							{props.name}: <span class={tw`text-cyan-200`}>{props.tag}</span>
 						</div>
-					</Motion.div>
-				</Show>
-			</Presence>
+						<div class={tw`font-mono text-sm leading-4 invisible`}>
+							{props.name}: {props.tag}
+						</div>
+					</div>
+				</div>
+			</Show>
 		</div>
 	)
 }